% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_types.R
\docType{data}
\name{allowed_vals_double}
\alias{allowed_vals_double}
\title{Allowed double-letter IUPAC}
\format{
A character vector with 17 elements
}
\usage{
allowed_vals_double
}
\description{
A character vector containing the valid double-letter IUPAC
}
\examples{
allowed_vals_double 

"AA" = Adenine, "CC" = Cytosine, "GG" = Guanine, "TT" = Thymine
"K" = GT|TG, "M" = AC|CA, "R" = AG|GA, "S" = GC|CG, "W" = AT|TA, "Y" = CT|TC
"--" = missing data

}
\keyword{datasets}
