% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_types.R
\docType{data}
\name{allowed_vals_single}
\alias{allowed_vals_single}
\title{Allowed single-letter IUPAC codes}
\format{
A character vector with 11 elements
}
\usage{
allowed_vals_single
}
\description{
A character vector containing all valid single-letter IUPAC nucleotide codes
}
\examples{
allowed_vals_single 

"A" = Adenine, "C" = Cytosine, "G" = Guanine, "T" = Thymine
"K" = G/T, "M" = A/C, "R" = A/G, "S" = G/C, "W" = A/T, "Y" = C/T
"-" = missing data

}
\keyword{datasets}
