% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assert}
\alias{assert}
\title{Function to assert a set of conditions}
\usage{
assert(statement, err_message = NULL)
}
\arguments{
\item{statement}{vector of comparison to proof}

\item{err_message}{string of error message when not all comparison are TRUE}
}
\value{
character string used as error message or nothing if all TRUE
}
\description{
Function to assert a set of conditions
}
\keyword{internal}
