% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calc_allelic_stats}
\alias{calc_allelic_stats}
\title{Analyse Allelic Status of SNPs}
\usage{
calc_allelic_stats(processed_snp_data)
}
\arguments{
\item{processed_snp_data}{SNP data frame that has been converted to single-letter IUPAC by read_raw_snp_data().}
}
\value{
A list containing:
\describe{
\item{}{A data frame the fraction of polymorphic, monomorphic, and failed SNP calls}
\item{}{The allele_df data frame that contains all essential marker metrics}
}
}
\description{
This function checks whether a SNP call is bi-allelic (polymorphic), monomorphic, or completely failed.
It summarizes allele counts for each SNP and calculates the frequencies.
It invokes the calc_polymorphism() function and generates the allele_df frame.
The allele_df data frame contains all essential information about each marker.
}
