% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_he.R
\name{calc_he}
\alias{calc_he}
\title{Determining the Heterozygosity Value}
\usage{
calc_he(df)
}
\arguments{
\item{allele_df}{Data frame that contains allele counts and frequencies for each marker.}
}
\value{
Adds an additional column to the allele data frame:
\describe{
\item{}{The column contains the He value for each marker in the allele data frame}
}
}
\description{
This function calculates the heterozygosity value (He) of a marker after Nei (1978).
}
