% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calc_polymorphism}
\alias{calc_polymorphism}
\title{Determining the Informativeness of a Marker}
\usage{
calc_polymorphism(df)
}
\arguments{
\item{allele_df}{Data frame that contains allele counts and frequencies for each marker.}
}
\value{
Adds two additional columns to the allele data frame:
\describe{
\item{}{The column contains the He and PIC value for each marker in the allele data frame}
}
}
\description{
This function calculates the heterozygosity value (He) of a marker after Nei (1978)
and the Polymorphism Information Content (PIC) of a marker after Botstein (1980).
}
