% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_snp_alleles.R
\name{calc_snp_alleles}
\alias{calc_snp_alleles}
\title{Analyse Allelic Status of SNPs}
\usage{
calc_snp_alleles(processed_snp_data)
}
\arguments{
\item{processed_snp_data}{SNP data frame that has been converted to single-letter IUPAC by read_raw_snp_data().}
}
\value{
A list containing:
\describe{
\item{}{A data frame the fraction of polymorphic, monomorphic, and failed SNP calls}
}
}
\description{
This function checks whether a SNP call is bi-allelic (polymorphic), monomorphic, or completely failed.
}
