% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{chars}
\alias{chars}
\title{Convert string to individual chars}
\usage{
chars(str)
}
\arguments{
\item{str}{A string to be converted into chars}
}
\value{
A vector of chars
}
\description{
Convert string to individual chars
}
\examples{
chars("ATGC")
# Returns: c("A", "T", "G", "C")
}
