% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_unrecognized}
\alias{check_unrecognized}
\title{Check for unrecognized SNP values}
\usage{
check_unrecognized(unique_vals, allowed_vals)
}
\arguments{
\item{unique_vals}{Vector of unique SNP values from the matrix}

\item{allowed_vals}{Vector of allowed characters}
}
\description{
Check for unrecognized SNP values
}
