% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{create_snphub_object}
\alias{create_snphub_object}
\title{Create a SNPhub S3 Object}
\usage{
create_snphub_object(
  snp_data,
  snp_summary,
  marker_stats = NULL,
  output_dir = NULL,
  marker_names = NULL,
  allele_df = NULL
)
}
\arguments{
\item{output_dir}{Output directory (optional).}

\item{raw_data}{Raw or processed SNP data frame.}

\item{summary}{SNP call summary statistics.}

\item{alleles}{Allele info (optional).}
}
\value{
An S3 object of class "snp_data"
}
\description{
Create a SNPhub S3 Object
}
