% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{filter_method}
\alias{filter_method}
\title{Calculates Nei's genetic diversity (expected heterozygosity, He) and the
Polymorphism Information Content (PIC) for each marker}
\usage{
filter_method(p, q, method)
}
\arguments{
\item{p}{Allele frequency of allele A}

\item{q}{Allele frequencey of allele B}

\item{method}{filter method to apply on given marker}
}
\value{
Numerical value of the informativeness of genetic marker based on the selected method
}
\description{
Calculates Nei's genetic diversity (expected heterozygosity, He) and the
Polymorphism Information Content (PIC) for each marker
}
