% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_types.R
\docType{data}
\name{iupac_map}
\alias{iupac_map}
\title{IUPAC conversion table (double-letter to single-letter)}
\format{
Named character vector of length 17
}
\usage{
iupac_map
}
\description{
An named character vector that maps valid double-letter IUPAC genotypes to
corresponding single-letter codes
}
\details{
Used during genotype standardization in 'read_raw_snp_data()'
}
\keyword{datasets}
