% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_marker_status}
\alias{plot_marker_status}
\title{Plot Marker Status}
\usage{
plot_marker_status(df, output, title)
}
\arguments{
\item{df}{The data frame containing the fraction of polymorphic, monomorphic, and failed marker}

\item{output}{Path to the output PNG file}

\item{title}{Title of the barplot}
}
\description{
Plot Marker Status
}
