% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_nei_pic_boxplot}
\alias{plot_nei_pic_boxplot}
\title{Plot Polymorphism Information Content}
\usage{
plot_nei_pic_boxplot(df, output, title)
}
\arguments{
\item{df}{The allele_df data frame from calc_allelic_status}

\item{output}{Path to the output PNG file}

\item{title}{Title of the boxplot}
}
\description{
Plot Polymorphism Information Content
}
