% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_snp_distribution}
\alias{plot_snp_distribution}
\title{Plot SNP Call Distribution}
\usage{
plot_snp_distribution(snp_counts, output, title)
}
\arguments{
\item{snp_counts}{A named list with SNP call counts}

\item{output}{Path to the output PNG file}

\item{title}{Title of the barplot}
}
\description{
Plot SNP Call Distribution
}
