% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remap_snps}
\alias{remap_snps}
\title{Function to replace SNPs values of an matrix by a named vector}
\usage{
remap_snps(mat, remap_list)
}
\arguments{
\item{mat}{original matrix}

\item{remap_list}{named vector of SNPs to replace the original matrix values with}
}
\description{
Function to replace SNPs values of an matrix by a named vector
}
\examples{
remap_snps(matrix(c("AA", "AG", "--", "GG"), ncol = 2), c("AA" = "A"))
# Returns: matrix(c("A", "AG", "--", "GG"), ncol = 2)
}
\keyword{internal}
