% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_snphub_object.R
\name{snp_summary}
\alias{snp_summary}
\title{Print Method for SNP Stats}
\usage{
snp_summary(x, ...)
}
\arguments{
\item{x}{The S3 objectcreated by read_raw_snp_data()}
}
\description{
Print Method for SNP Stats
}
