% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{snphub_summary}
\alias{snphub_summary}
\title{Print Method for SNP Stats}
\usage{
snphub_summary(x, ...)
}
\arguments{
\item{x}{The S3 objectcreated by read_raw_snp_data()}
}
\description{
Print Method for SNP Stats
}
