% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calc_percent}
\alias{calc_percent}
\title{Calculate percentage}
\usage{
calc_percent(count, total)
}
\arguments{
\item{count}{Numerical value}

\item{total}{Maximum value}
}
\description{
Calculate percentage
}
