% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calc_polymorphism}
\alias{calc_polymorphism}
\title{Determining the Informativeness of a Marker}
\usage{
calc_polymorphism(df)
}
\arguments{
\item{allele_df}{Data frame with allele count columns and frequencies for each marker}
}
\value{
A modified version of the input data frame with three new columns:
\itemize{
\item \code{Nei_H}:Nei's genetic diveristy (expected heterozgosity)
\item \code{PIC}: Polymorphism Information Content
\item \code{MAF}: Minor Allele Frequency (only for polymorphic markers)
}
}
\description{
Calculates Nei's genetic diversity (expected heterozygosity, He) and the
Polymorphism Information Content (PIC) for each marker
}
