% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{chars}
\alias{chars}
\title{Convert string to individual characters}
\usage{
chars(str)
}
\arguments{
\item{str}{A string to be converted into chars}
}
\value{
A vector of characters

single letter of string
}
\description{
Convert string to individual characters
}
\keyword{internal}
